requirements.txt
AccessControl==6.3
Acquisition==5.2
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.1
altgraph==0.17.3
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.5.0
ansi2html==1.8.0
anyio==4.1.0
apispec==6.4.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async==0.6.2
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
AuthEncoding==5.0
auto-py-to-exe==2.30.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
billiard==4.2.0
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
BTrees==5.2
cachelib==0.12.0
cachetools==5.3.1
catalogue==2.0.10
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
Chameleon==4.4.3
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.0
confection==0.1.3
configparser==6.0.0
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
cymem==2.0.8
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datatables==0.4.9
dateparser==1.1.8
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
DocumentTemplate==4.6
docutils==0.20.1
dominate==2.9.1
Eel==0.14.0
email-validator==1.1.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
executing==1.2.0
ExtensionClass==5.1
Faker==0.7.12
fastjsonschema==2.19.0
filelock==3.14.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Assets==2.1.0
flask-babel==4.0.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Celery==2.4.3
Flask-CeleryExt==0.5.0
Flask-Cors==4.0.0
Flask-DebugToolbar==0.15.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-ReCaptcha==0.4.2
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
Flask-User==1.0.2.2
Flask-WAT==0.1.3
Flask-WTF==1.2.1
Flask-WTF-Ext==0.2
Flask-WTF-FlexWidgets==0.1.25
fonttools==4.43.1
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.4.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
git-filter-repo==2.38.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-speech==2.21.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna==3.4
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isbnlib==3.10.3
iso3166==1.0.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kaleido==0.2.1
keyboard==0.13.5
kiwisolver==1.4.5
kombu==5.3.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
limits==3.7.0
llvmlite==0.41.0
lxml==4.9.3
macholib==1.16.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.7
msgspec==0.18.6
multidict==6.0.4
MultiMapping==5.0
multipart==0.2.4
multitasking==0.0.11
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.0
numpy==1.25.2
openai==0.28.0
optional-django==0.3.0
ordered-set==4.1.0
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
passlib==1.7.4
PasteDeploy==3.1.0
pathy==0.10.2
Persistence==4.1
persistent==5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipdeptree==2.13.2
pipenv==2023.11.17
pipreqs==0.5.0
platformdirs==3.11.0
plotly==5.17.0
pocketsphinx==5.0.2
pooch==1.7.0
preshed==3.0.9
prison==0.2.1
prometheus-client==0.19.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-moneyed==0.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==23.12.11
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.0.9
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-gettext==5.0
python-json-logger==2.0.7
pytz==2023.3.post1
pyvis==0.3.2
pyvista==0.42.3
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
querystring_parser==1.2.3
redis==5.0.1
referencing==0.32.0
regex==2023.8.8
requests==2.31.0
RestrictedPython==7.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roman==4.1
rpds-py==0.13.2
rsa==4.9
schwifty==2020.9.0
scikit-learn==1.3.1
scipy==1.11.3
scooby==0.7.4
selenium==4.13.0
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.6
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speaklater==1.3
SpeechRecognition==3.10.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
squarify==0.4.3
srsly==2.4.8
stack-data==0.6.2
stripe>=10.0.0
tenacity==8.2.3
terminado==0.18.0
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.10.0
transaction==4.0
trio==0.22.2
trio-websocket==0.11.1
twilio==9.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.14
vine==5.1.0
virtualenv==20.26.2
visitor==0.1.3
vtk==9.2.6
waitress==3.0.0
wasabi==1.1.2
wcwidth==0.2.6
webassets==2.0
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
webpack==6.0.0
websocket-client==1.7.0
WebTest==3.0.0
Werkzeug==2.2.3
whichcraft==0.6.1
widgetsnbextension==4.0.9
wrapt==1.16.0
WSGIProxy2==0.5.1
wsproto==1.2.0
WTForms==3.1.1
WTForms-Ext==0.5
WTForms-SQLAlchemy==0.4.1
wtforms_extras==1.1.4
yarg==0.1.9
yarl==1.9.2
z3c.pt==4.3
zc.lockfile==3.0.post1
ZConfig==4.0
zExceptions==5.0
zipp==3.17.0
ZODB==5.8.1
zodbpickle==3.2
Zope==5.9
zope.annotation==5.0
zope.browser==3.0
zope.browsermenu==5.0
zope.browserpage==5.0
zope.browserresource==5.1
zope.cachedescriptors==5.0
zope.component==6.0
zope.configuration==5.0.1
zope.container==5.2
zope.contentprovider==5.0
zope.contenttype==5.1
zope.datetime==5.0.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.dottedname==6.0
zope.event==4.6
zope.exceptions==5.0.1
zope.filerepresentation==6.0
zope.globalrequest==2.0
zope.hookable==6.0
zope.i18n==5.1
zope.i18nmessageid==6.1.0
zope.interface==5.5.2
zope.lifecycleevent==5.0
zope.location==5.0
zope.pagetemplate==5.1
zope.processlifetime==3.0
zope.proxy==5.2
zope.ptresource==5.0
zope.publisher==7.0
zope.schema==7.0.1
zope.security==6.2
zope.sequencesort==5.0
zope.site==5.0
zope.size==5.0
zope.structuredtext==5.0
zope.tal==5.0.1
zope.tales==6.0
zope.testbrowser==6.0
zope.testing==5.0.1
zope.traversing==5.0
zope.viewlet==5.0
